Move StorageKeys from types -> constants
